Add files via upload
The output of this function, M, is a 3D array:  the number of magnetization indexes through the slice's direction, 3 (Mx, My, and Mz), and the number timepoints in the excitation's pulse.  As such, the output gives the magnetization vector at each timepoint within the RF's excitation pulse for each location along the slice's direction.  This function calculates the excitation gradient appropriate to achieve the slice-thickness with the inputted excitation pulse.  It also gives the RF's pulse the appropriate phase-ramp needed for off-isocenter excitation. 

The inputs are the following:
1. B1e (RF's envelope function for each timepoint.  Each unit is in Tesla).
2. sliceDir (The location, in mm, of each magnetization vector along the slice's direction). 
3. RFBW (The RF's excitation bandwidth, in Hz).
4. sliceThickness (The nominal slice thickness of the RF pulse.  gamma * slice select gradient * sliceThickness will be the full width at half-maximum of the excitation's profile). 
5. RFDur (The duration of the Rf's excitation in seconds)
6. M0 (An array consisting of each initialized magnetization vector along space with dimensions length(sliceDir) x 3). 
7. gamma (the Gyromagnetic ratio in rad/S/T.
8.  sliceOffCenter (the mm length off of isocenter for the center target of the excitation pulse). 
9. RFPhase0 (the initial RF's phase). 
10. Meq (the equilibirium magnetization for each magnetization vector.  dimensions length(sliceDir) x 1).
11. T1, T2:  longitudinal and transverse recovery/decay constants.  I made it the same for all spins.
12. dfArr (the local off-resonance for each spin in Hz.  of dimensions length(sliceDir x 1)